+++ misc/client-extensions/CMakeLists.txt
@@ -15,4 +15,3 @@ target_compile_options(lua PRIVATE -w)
 add_subdirectory(CustomPackets)
 add_subdirectory(tests)
 add_subdirectory(ClientExtensions)
-add_subdirectory(CustomDBCMgr)
