+++ misc/client-extensions/ClientExtensions/ClientDB.cpp
@@ -1,11 +1,6 @@
 #include "ClientDB.h"
 
-void ClientDB::Apply() {
-    OverwriteUInt32AtAddress(0x634E30, CalculateAddress(reinterpret_cast<uint32_t>(&RegisterBaseEx), 0x634E30));
+void ClientDB::Load() {
+
 }
 
-uint32_t ClientDB::RegisterBaseEx(uint32_t a) {
-    uint32_t result = ClientDB__RegisterBase(a);
-
-    return result;
-}
