+++ misc/client-extensions/ClientExtensions/SharedDefines.h
@@ -60,10 +60,19 @@ static char* sPluralS = "s";
 static char* sSpace = " ";
 
 // client functions
+// Defs cherrypicked from StormLib: https://github.com/ladislav-zezula/StormLib
+CLIENT_FUNCTION(SFileOpenFileEx, 0x424B50, __stdcall, bool, (HANDLE, const char*, uint32_t, HANDLE*))
+CLIENT_FUNCTION(SFileReadFile, 0x422530, __stdcall, bool, (HANDLE handle /*likely a handle*/, void* data, uint32_t bytesToRead, uint32_t* bytesRead, uint32_t* overlap /*just set to 0*/))
+CLIENT_FUNCTION(SFileCloseFile, 0x422910, __stdcall, void, (HANDLE a1))
+
+//
 CLIENT_FUNCTION(SFileOpenFile, 0x424F80, __stdcall, int, (char const* filename, HANDLE* a2 /*file handle out*/))
 CLIENT_FUNCTION(SFileGetFileSize, 0x4218C0, __stdcall, DWORD /*lowest 32 bits in size*/, (HANDLE handle, DWORD* highSize /*highest 32 bits in size*/))
-CLIENT_FUNCTION(SFileReadFile, 0x422530, __stdcall, int, (HANDLE handle /*likely a handle*/, char* data, DWORD bytesToRead, DWORD* bytesRead, int overlap /*just set to 0*/, int unk /*just set to 0*/))
-CLIENT_FUNCTION(SFileCloseFile, 0x422910, __stdcall, int, (HANDLE a1))
+
+CLIENT_FUNCTION(SMemAlloc, 0x76E540, __stdcall, void*, (uint32_t, const char*, uint32_t, uint32_t))
+CLIENT_FUNCTION(SMemFree, 0x76E5A0, __stdcall, bool, (void*, const char*, uint32_t, uint32_t))
+
+CLIENT_FUNCTION(SErrPrepareAppFatal, 0x772A80, _cdecl, void, (uint32_t, const char*, ...))
 
 CLIENT_FUNCTION(ClntObjMgrGetActivePlayer, 0x4D3790, __cdecl, unsigned long long, ())
 CLIENT_FUNCTION(ClntObjMgrObjectPtr, 0x4D4DB0, __cdecl, void*, (unsigned long long, unsigned int))
